Fix compile error

+++ client/src/ChunkMesher.cpp
@@ -69,10 +69,10 @@ void ChunkMesher::Run() {
 		}
 
 	if (m_init_light) {
-		using LightAlgo =
-		    BlockLightAlgo<BlockAlgoConfig<int32_t, BlockAlgoBound<int32_t>(-1, -1, -1, kChunkSize + 1, kChunkSize + 1,
-		                                                                    kChunkSize + 1)>,
-		                   14>;
+		using LightAlgo = BlockLightAlgo<
+		    BlockAlgoConfig<int32_t, BlockAlgoBound<int32_t>(-1, -1, -1, (int32_t)kChunkSize + 1,
+		                                                     (int32_t)kChunkSize + 1, (int32_t)kChunkSize + 1)>,
+		    14>;
 
 		LightAlgo::Queue sunlight_entries, torchlight_entries;
 		for (int32_t y = -15; y < (int32_t)kChunkSize + 15; ++y)
